+++ package.json
@@ -39,6 +39,7 @@
     "react-rating-stars-component": "^2.2.0",
     "react-router-dom": "^6.15.0",
     "react-scripts": "5.0.1",
+    "react-toastify": "^9.1.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
