+++ src/Component/Common/i18/index.jsx
@@ -1,6 +1,5 @@
 import i18n from "i18next";
 import { initReactI18next } from "react-i18next";
-
 import Backend from "i18next-http-backend";
 import LanguageDetector from "i18next-browser-languagedetector";
