+++ src/Component/language/index.jsx
@@ -1,13 +1,16 @@
 import React from "react";
-import lgimg from "../../assets/img/logo.jpg";
-import "../../assets/css/lang.css";
 import { Link } from "react-router-dom";
-import Select from "../Common/Select";
-// import { useTranslation } from "react-i18next";
+
+// Third party i18next
 import i18next from "i18next";
 
-const Lang = () => {
-  // const { t } = useTranslation();
+// Common Component
+import Select from "../Common/Select";
+
+// Image
+import lgimg from "../../assets/img/logo.jpg";
+
+const Language = () => {
   const handleChangen = (e) => {
     i18next.changeLanguage(e.target.value);
   };
@@ -27,9 +30,8 @@ const Lang = () => {
           <li></li>
         </ul>
       </div>
-
       <section className="container-fluid context">
-        <div className=" row justify-content-center ">
+        <div className="row justify-content-center ">
           <div className=" lang-box col-lg-5 d-flex justify-content-around py-3">
             <div className="row">
               <div className="col-lg-6">
@@ -38,7 +40,6 @@ const Lang = () => {
                 </div>
               </div>
               <div className="col-lg-6">
-
                 <div className="lang-box2 ">
                   <Select onChange={(e) => handleChangen(e)} />
                   <div className="lang-link">
@@ -56,4 +57,4 @@ const Lang = () => {
   );
 };
 
-export default Lang;
+export default Language;
