+++ src/index.js
@@ -6,7 +6,7 @@ import reportWebVitals from "./reportWebVitals";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "bootstrap/dist/js/bootstrap.bundle.min";
 import { BrowserRouter } from "react-router-dom";
-import './Component/i18/index';
+import './Component/Common/i18/index';
 
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
